Moved the common code up the superclass (#171)
