+++ dz3-swing/src/main/java/net/sf/dz3/view/swing/thermostat/Chart2016.java
@@ -278,72 +278,4 @@ public class Chart2016 extends AbstractChart {
             }
         }
     }
-
-    /**
-     * Averaging tool.
-     */
-    private class Averager {
-
-        /**
-         * The expiration interval. Values older than the last key by this many
-         * milliseconds are expired.
-         */
-        private final long expirationInterval;
-
-        /**
-         * The timestamp of the oldest recorded sample.
-         */
-        private Long timestamp;
-
-        private int count;
-        private double valueAccumulator = 0;
-        private double tintAccumulator = 0;
-        private double emphasizeAccumulator = 0;
-
-        public Averager(long expirationInterval) {
-            this.expirationInterval = expirationInterval;
-        }
-
-        /**
-         * Record a value.
-         *
-         * @param signal Signal to record.
-         *
-         * @return The average of all data stored in the buffer if this sample is more than {@link #expirationInterval}
-         * away from the first sample stored, {@code null} otherwise.
-         */
-        @SuppressWarnings("squid:S2629")
-        public TintedValue append(DataSample<? extends TintedValue> signal) {
-
-            if (timestamp == null) {
-                timestamp = signal.timestamp;
-            }
-
-            long age = signal.timestamp - timestamp;
-
-            if ( age < expirationInterval) {
-
-                count++;
-                valueAccumulator += signal.sample.value;
-                tintAccumulator += signal.sample.tint;
-                emphasizeAccumulator += signal.sample.emphasize ? 1 : 0;
-
-                return null;
-            }
-
-            // VT: NOTE: squid:S2629 - this happens once in 25-40 seconds, acceptable loss
-
-            //logger.debug("RingBuffer: flushing at {}", Interval.toTimeInterval(age));
-
-            TintedValue result = new TintedValue(valueAccumulator / count, tintAccumulator / count, emphasizeAccumulator > 0);
-
-            count = 1;
-            valueAccumulator = signal.sample.value;
-            tintAccumulator = signal.sample.tint;
-            emphasizeAccumulator = 0;
-            timestamp = signal.timestamp;
-
-            return result;
-        }
-    }
 }
